feat: move i18n config in config.json
