--- ui/public/i18n.json
@@ -1,4 +1,0 @@
-{
-  "languages": ["fr-FR", "en-US"],
-  "locale": "fr-FR"
-}
